+++ src/main/java/com/macuguita/lib/fabric/FabricPlatformImpl.java
@@ -22,7 +22,8 @@ package com.macuguita.lib.fabric;
 
 //? fabric {
 
-import java.nio.file.Path;
+/*import java.nio.file.Path;
+import java.util.function.Consumer;
 
 import com.macuguita.lib.Platform;
 import com.macuguita.lib.fabric.network.FabricClientNetworkBootstrap;
@@ -83,7 +84,7 @@ public class FabricPlatformImpl implements Platform {
 	public <T extends CustomPacketPayload> void registerC2S(
 			NetworkManager.C2SRegistration<T> reg
 	) {
-		PayloadTypeRegistry./*? >= 26.1 {*/serverboundPlay/*?} else {*//*playC2S*//*?}*/().register(reg.type(), reg.codec());
+		PayloadTypeRegistry./^? >= 26.1 {^/serverboundPlay/^?} else {^//^playC2S^//^?}^/().register(reg.type(), reg.codec());
 		ServerPlayNetworking.registerGlobalReceiver(
 				reg.type(),
 				(payload, context) -> {
@@ -97,8 +98,9 @@ public class FabricPlatformImpl implements Platform {
 	public <T extends CustomPacketPayload> void registerS2C(
 			NetworkManager.S2CRegistration<T> reg
 	) {
-		PayloadTypeRegistry./*? >= 26.1 {*/clientboundPlay/*?} else {*//*playS2C*//*?}*/().register(reg.type(), reg.codec());
-		if (FabricLoader.getInstance().getEnvironmentType() == EnvType.CLIENT) { // This should be safe because it only returns true when it is on the client jar
+		PayloadTypeRegistry./^? >= 26.1 {^/clientboundPlay/^?} else {^//^playS2C^//^?}^/().register(reg.type(), reg.codec());
+		if (FabricLoader.getInstance().getEnvironmentType() == EnvType.CLIENT // This should be safe because it only returns true when it is on the client jar
+				&& reg.handlerSupplier() != null) {
 			FabricClientNetworkBootstrap.registerS2CHandler(
 					reg.type(),
 					(payload, context) -> {
@@ -108,5 +110,16 @@ public class FabricPlatformImpl implements Platform {
 			);
 		}
 	}
+
+	@Override
+	public <T extends CustomPacketPayload> void registerClientS2CHandler(
+			CustomPacketPayload.Type<T> type,
+			Consumer<T> handler
+	) {
+		ClientPlayNetworking.registerGlobalReceiver(
+				type,
+				(payload, context) -> handler.accept(payload)
+		);
+	}
 }
-//?}
+*///?}