+++ src/main/java/com/macuguita/lib/fabric/network/FabricClientNetworkBootstrap.java
@@ -22,7 +22,7 @@ package com.macuguita.lib.fabric.network;
 
 //? fabric {
 
-import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+/*import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 
@@ -35,4 +35,4 @@ public class FabricClientNetworkBootstrap {
 		);
 	}
 }
-//?}
+*///?}