+++ src/testmod/java/com/macuguita/libtest/fabric/FabricClientEntrypoint.java
@@ -21,7 +21,7 @@
 package com.macuguita.libtest.fabric;
 
 //? fabric {
-import com.macuguita.libtest.client.TestModClient;
+/*import com.macuguita.libtest.client.TestModClient;
 import net.fabricmc.api.ClientModInitializer;
 
 public class FabricClientEntrypoint implements ClientModInitializer {
@@ -30,4 +30,4 @@ public class FabricClientEntrypoint implements ClientModInitializer {
         TestModClient.init();
     }
 }
-//?}
+*///?}