+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-2.0.0-SNAPSHOT
+2.0.1
+- Expose kaleido config with api gradle config
+- Allow to register packets without handlers
+---
+2.0.0
 - Made the supporter cape system better
 - Update to latest minecraft version
 - Simplified the api